Create a topic for publishing a scene with new entities.
Uses EachNew to create a scene message consisting of only new entities. This changeset also introduces a change in the content of the scene graph in `SceneBroadcaster`.  The data in each node is now the message associated with that entity only. i.e, a model node only has a message only about the model. It will not have any links, joints, etc. To create a the whole scene, one has to traverse the whole graph adding messages as necessary.